refactor: remove unnecessary logs + create global style file + combine related interfaces in one file Co-authored-by: nabil.salah203@gmail.com <nabil.salah203@gmail.com>
