+++ frontend/public/global.css
@@ -1,0 +1,21 @@
+.background-green {
+  background-color: #1aa18f !important;
+}
+.thick-border .v-data-table__wrapper {
+  border: 3px solid #000;
+}
+.v-data-table-footer__items-per-page { 
+  display: none !important;  
+}
+.v-data-table td{
+  padding: 4px 8px;
+  font-size: 12px;
+  font-weight: 500;
+}
+.mn-height {
+  min-height: calc(100% - 7%);
+}
+
+.file-name {
+  font-weight: 500;
+}
