--- frontend/public/theme.css
@@ -1,3 +1,0 @@
-.background-green {
-  background-color: #1aa18f !important;
-}
